Consolidate signature reading and writing into one file and namespace
