+++ src/library/impl/meta_reader/cache.h
@@ -1,3 +1,4 @@
+#pragma once
 
 namespace xlang::meta::reader
 {
