+++ src/library/impl/meta_reader/database.h
@@ -1,3 +1,4 @@
+#pragma once
 
 namespace xlang::impl
 {
