+++ src/library/impl/meta_reader/enum_traits.h
@@ -1,3 +1,4 @@
+#pragma once
 
 namespace xlang::meta::reader
 {
