+++ src/library/impl/meta_reader/flags.h
@@ -1,3 +1,5 @@
+#pragma once
+
 namespace xlang::impl
 {
     template <typename T>
