+++ src/library/impl/meta_reader/helpers.h
@@ -1,18 +1,7 @@
+#pragma once
 
 namespace xlang::meta::reader
 {
-    template <typename T>
-    bool empty(std::pair<T, T> const& range) noexcept
-    {
-        return range.first == range.second;
-    }
-
-    template <typename T>
-    std::size_t size(std::pair<T, T> const& range) noexcept
-    {
-        return range.second - range.first;
-    }
-
     inline auto find(TypeRef const& type)
     {
         return type.get_database().get_cache().find(type.TypeNamespace(), type.TypeName());
@@ -40,7 +29,7 @@ namespace xlang::meta::reader
         }
     }
 
-    inline bool is_const(ParamSig const& param)
+    inline bool is_const(meta::signature::ParamSig const& param)
     {
         auto is_type_const = [](auto&& type)
         {
