+++ src/library/meta_writer.h
@@ -1,4 +1,4 @@
 #pragma once
 
 #include "impl/meta_writer/pe_writer.h"
-#include "impl/meta_writer/signature_writer.h"
+#include "impl/signature.h"
